requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==4.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
importlib_metadata==8.5.0
Markdown==3.7
psycopg2-binary==2.9.10
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.2.3
zipp==3.20.2
